Remove side effects from wbplanview action reducer
Fixes #882
