+++ specifyweb/frontend/js_src/lib/wbplanviewloadingreducer.ts
@@ -1,17 +1,29 @@
+import $ from 'jquery';
 import type { State } from 'typesafe-reducer';
 import { generateDispatch } from 'typesafe-reducer';
 
-import type { SpecifyResource } from './components/wbplanview';
-import navigation from './navigation';
+import type {
+  SpecifyResource,
+  WbPlanViewWrapperProps,
+} from './components/wbplanview';
+import type { MappingState } from './components/wbplanviewstate';
+import { mappingsTreeToUploadPlan } from './mappingstreetouploadplan';
+import { renameNewlyCreatedHeaders } from './wbplanviewheaderhelper';
+import { getMappingsTree, getMustMatchTables, goBack } from './wbplanviewutils';
 
 type NavigateBackState = State<
   'NavigateBackState',
   {
-    readonly wb: SpecifyResource;
+    wb: SpecifyResource;
   }
 >;
 
-export type LoadingStates = NavigateBackState;
+type SavePlanState = State<
+  'SavePlanState',
+  { props: WbPlanViewWrapperProps; state: MappingState }
+>;
+
+export type LoadingStates = NavigateBackState | SavePlanState;
 
 export const loadingStateDispatch = generateDispatch<LoadingStates>({
   NavigateBackState: (state): void =>
@@ -19,5 +31,64 @@ export const loadingStateDispatch = generateDispatch<LoadingStates>({
      * Need to make the `Loading` dialog
      * appear before the `Leave Page?` dialog
      */
-    void setTimeout(() => navigation.go(`/workbench/${state.wb.id}/`), 10),
+    void setTimeout(() => goBack(state.wb.id), 10),
+  SavePlanState({ state, props }): void {
+    const renamedMappedLines = renameNewlyCreatedHeaders(
+      state.baseTableName,
+      props.dataset.columns,
+      state.lines
+    );
+
+    const newlyAddedHeaders = renamedMappedLines
+      .filter(
+        ({ headerName, mappingPath }) =>
+          mappingPath.length > 0 &&
+          mappingPath[0] !== '0' &&
+          !props.dataset.columns.includes(headerName)
+      )
+      .map(({ headerName }) => headerName);
+
+    const uploadPlan = mappingsTreeToUploadPlan(
+      state.baseTableName,
+      getMappingsTree(renamedMappedLines, true),
+      getMustMatchTables(state)
+    );
+
+    const dataSetRequestUrl = `/api/workbench/dataset/${props.dataset.id}/`;
+
+    void $.ajax(dataSetRequestUrl, {
+      type: 'PUT',
+      data: JSON.stringify({
+        uploadplan: uploadPlan,
+      }),
+      dataType: 'json',
+      processData: false,
+    }).done(() => {
+      if (state.changesMade) props.removeUnloadProtect();
+
+      if (newlyAddedHeaders.length > 0)
+        $.ajax(dataSetRequestUrl, {
+          type: 'GET',
+        }).done(({ columns, visualorder }) => {
+          const newVisualOrder =
+            visualorder === null
+              ? Object.keys(props.dataset.columns)
+              : visualorder;
+
+          newlyAddedHeaders.forEach((headerName) =>
+            newVisualOrder.push(columns.indexOf(headerName))
+          );
+
+          $.ajax(dataSetRequestUrl, {
+            type: 'PUT',
+            data: JSON.stringify({
+              visualorder: newVisualOrder,
+            }),
+            dataType: 'json',
+            processData: false,
+          }).done(() => goBack(props.dataset.id));
+        });
+      else goBack(props.dataset.id);
+    });
+  },
 });
